+++ buildSrc/src/main/kotlin/ProjectExtensions.kt
@@ -3,6 +3,8 @@ import org.gradle.api.file.RegularFile
 import org.gradle.api.internal.project.ProjectInternal
 import org.gradle.jvm.toolchain.JavaLanguageVersion
 import org.gradle.jvm.toolchain.JavaToolchainService
+import kotlin.io.path.Path
+import kotlin.io.path.readText
 
 fun Project.getProperty(key: String) =
     providers.gradleProperty(key).get()
@@ -21,6 +23,21 @@ val Project.projectUrl: String
 
 val Project.projectDescription: String
     get() = getProperty("projectDescription")
+
+val Project.projectCommit: String
+    get() {
+        try {
+            val gitHead = Path(rootDir.path, ".git", "HEAD").readText()
+            if (gitHead.startsWith("ref: ")) {
+                return Path(rootDir.path, ".git", gitHead.substring(5).trim()).readText().trim()
+            } else {
+                return gitHead.trim()
+            }
+        } catch (e: Exception) {
+            logger.warn("Failed to read Git commit hash: {}", e.toString())
+        }
+        return "Unknown"
+    }
 
 val Project.isSnapshot: Boolean
     get() = projectVersion.endsWith("-SNAPSHOT")