+++ src/main/java/pascal/taie/analysis/pta/core/cs/element/Pointer.java
@@ -98,6 +98,16 @@ public interface Pointer extends Indexable {
     PointerFlowEdge addEdge(PointerFlowEdge edge);
 
     /**
+     * Removes out edges of this pointer if they satisfy the filter.
+     * <p>
+     * <strong>Note:</strong> This method should not be called outside of
+     * {@link pascal.taie.analysis.pta.plugin.Plugin#onPhaseFinish()},
+     * otherwise it may break the monotonicity of pointer analysis.
+     * </p>
+     */
+    void removeEdgesIf(Predicate<PointerFlowEdge> filter);
+
+    /**
      * @return out edges of this pointer in pointer flow graph.
      */
     Set<PointerFlowEdge> getOutEdges();