+++ src/main/java/pascal/taie/ir/exp/DoubleLiteral.java
@@ -22,7 +22,7 @@
 
 package pascal.taie.ir.exp;
 
-import pascal.taie.language.type.PrimitiveType;
+import pascal.taie.language.type.DoubleType;
 
 public class DoubleLiteral implements FloatingPointLiteral {
 
@@ -42,8 +42,8 @@ public class DoubleLiteral implements FloatingPointLiteral {
     }
 
     @Override
-    public PrimitiveType getType() {
-        return PrimitiveType.DOUBLE;
+    public DoubleType getType() {
+        return DoubleType.DOUBLE;
     }
 
     public double getValue() {