+++ src/main/java/pascal/taie/language/type/TypeSystem.java
@@ -28,8 +28,8 @@ import java.io.Serializable;
 
 /**
  * This class provides APIs for retrieving types in the analyzed program.
- * For convenience, the special predefined types, i.e., primitive types,
- * null type, and void type can be directly retrieved from their own classes.
+ * For convenience, null type, void type and single primitive type
+ * can be directly retrieved from their own classes.
  */
 public interface TypeSystem extends Serializable {
 
@@ -43,9 +43,13 @@ public interface TypeSystem extends Serializable {
 
     ArrayType getArrayType(Type baseType, int dimensions);
 
+    PrimitiveType getPrimitiveType(String typeName);
+
     ClassType getBoxedType(PrimitiveType type);
 
     PrimitiveType getUnboxedType(ClassType type);
 
     boolean isSubtype(Type supertype, Type subtype);
+
+    boolean isPrimitiveType(String typeName);
 }